add login and register

+++ app/Http/Controllers/TasksController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 
 use App\Http\Requests;
 use App\Task;
+use App\User;
 
 class TasksController extends Controller
 {
@@ -16,11 +17,20 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        return view('tasks.index', [
-            'tasks' => $tasks
-            ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            $data += $this->counts($user);
+            return view('tasks.index', $data);
+        } else {
+            return view('welcome');
+        }
     }
 
     /**
@@ -50,14 +60,14 @@ class TasksController extends Controller
             'status' => 'required|max:10',
         ]);
 
-        $task = new Task;
 
-        $task->content = $request->content;
-        $task->status = $request->status;
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
 
-        $task->save();
+        ]);
 
-        return redirect('/');
+        return redirect('/tasks/');
     }
 
     /**
@@ -70,10 +80,14 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
 
-        return view(
-            'tasks.show',
-            ['task'=>$task]
-        );
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.show', [
+                'task' => $task
+            ]);
+            return redirect('/tasks/');
+        } else {
+            return redirect('/');
+        }
     }
 
     /**
@@ -85,10 +99,15 @@ class TasksController extends Controller
     public function edit($id)
     {
         $task = Task::find($id);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', [
+                'task' => $task
+            ]);
 
-        return view('tasks.edit', [
-        'task' => $task
-        ]);
+            return redirect('/tasks/');
+        } else {
+            return redirect('/');
+        }
     }
 
     /**
@@ -106,12 +125,10 @@ class TasksController extends Controller
         ]);
 
         $task = Task::find($id);
-        $task->content = $request->content;
-        $task->status = $request->status;
 
-        $task->save();
+        $task->fill($request->all())->save();
 
-        return redirect('/');
+        return redirect('/tasks/');
     }
 
     /**
@@ -123,8 +140,12 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
 
-        return redirect('/');
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+            return redirect('/tasks/');
+        } else {
+            return redirect('/');
+        }
     }
 }
